BRIANS-SECOND-BRANCH-FOR-ADDING-LOMBOK added lombok and I did violate a principle because I refactored the current class packages and it should be in another commit
