+++ gradle.properties
@@ -1,0 +1,3 @@
+springPlatformVersion=Brussels-SR7
+springBootVersion=2.0.0.RELEASE
+lombokVersion=1.16.20
