+++ src/main/java/com/zoox/sample/git/Application.java
@@ -1,17 +1,18 @@
-package hello;
+package com.zoox.sample.git;
 
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
+
+import com.zoox.sample.git.domain.Customer;
+import com.zoox.sample.git.repository.CustomerRepository;
+import lombok.extern.slf4j.Slf4j;
 
 import org.springframework.boot.CommandLineRunner;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.context.annotation.Bean;
 
+@Slf4j
 @SpringBootApplication
 public class Application {
-
-	private static final Logger log = LoggerFactory.getLogger(Application.class);
 
 	public static void main(String[] args) {
 		SpringApplication.run(Application.class);
