+++ src/main/java/com/zoox/sample/git/repository/CustomerRepository.java
@@ -1,8 +1,10 @@
-package hello;
+package com.zoox.sample.git.repository;
 
 import java.util.List;
 
+import com.zoox.sample.git.domain.Customer;
 import org.springframework.data.repository.CrudRepository;
+
 
 public interface CustomerRepository extends CrudRepository<Customer, Long> {
